created base architecture with file system
