+++ config/index.js
@@ -7,8 +7,8 @@ const connectionString = `postgresql://${process.env.DB_USER}:${process.env.DB_P
 
 const pool = new Pool({
   connectionString: isProduction ? process.env.DATABASE_URL : connectionString,
-  // ssl: isProduction, // for local work
-  ssl: { rejectUnauthorized: false }, // for hosting
+  ssl: isProduction, // for local
+  // ssl: { rejectUnauthorized: false }, // for hosting
 })
 
 module.exports = {pool}
